--- src/api/cart.ts
@@ -1,16 +1,0 @@
-import axios from 'axios';
-import { ProductType } from './product';
-
-export interface CartItemType {
-  id: number;
-  product: ProductType;
-}
-export const fetchCartList = async (): Promise<null | CartItemType[]> => {
-  try {
-    const { data } = await axios(`${process.env.REACT_APP_API_PATH}/carts`);
-    return data.response;
-  } catch (error) {
-    console.error(error);
-    return null;
-  }
-};
